+++ src/mappings/BoolMapping.h
@@ -39,8 +39,8 @@ public:
      * @param readRestricted Is the Mapping write only?
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
-    BoolMapping(const std::string& reference, RegisterType registerType, int16_t address, bool readRestricted = false,
-                int8_t slaveAddress = -1);
+    BoolMapping(const std::string& reference, RegisterType registerType, int32_t address, bool readRestricted = false,
+                int16_t slaveAddress = -1);
 
     /**
      * @brief Constructor for TAKE_BIT type of Mappings.
@@ -52,8 +52,8 @@ public:
      * @param readRestricted Is the Mapping write only?
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
-    BoolMapping(const std::string& reference, RegisterType registerType, int16_t address, OperationType operation,
-                int8_t bitIndex, bool readRestricted = false, int8_t slaveAddress = -1);
+    BoolMapping(const std::string& reference, RegisterType registerType, int32_t address, OperationType operation,
+                int8_t bitIndex, bool readRestricted = false, int16_t slaveAddress = -1);
 
     /**
      * @brief Triggers the client to write the value for this register.
