+++ src/mappings/Int32Mapping.h
@@ -40,7 +40,7 @@ public:
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
     Int32Mapping(const std::string& reference, RegisterType registerType, const std::vector<int16_t>& addresses,
-                 OperationType operation, bool readRestricted = false, int8_t slaveAddress = -1);
+                 OperationType operation, bool readRestricted = false, int16_t slaveAddress = -1);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
